fix: Handle unbound variable error in associative array check
Use parameter expansion with default value (:-) to avoid 'unbound variable'
error when checking if a key exists in the seen_paths associative array.

This fixes the error:
  seen_paths[$target]: unbound variable

The script runs with set -u which treats unbound variables as errors,
so we need to use ${seen_paths[$target]:-} to provide an empty default.